postman: implement fast uploads
Previous behavior was "take user file, encode it to base64 and send back
to the user with captcha", so file was transferred 3 times!

As Sfor described, it was done intentially for slow down image-spam
posting. However, it was very annoying to i2p users because images
becomes corruptes often.

Now we're saving upload to in-memory object for 1 hour.